requirements.txt
tensorflow
pandas
dvc
mlflow
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
scipy
Flask
Flask-Cors
gdown
ipykernel
dagshub
platformdirs>=3.1.1,<4
-e .